Back_end/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0
python-dotenv>=1.0.0

# LangChain
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0

# Vector store
faiss-cpu>=1.8.0

# PDF parsing
pypdf>=4.0.0

# Embeddings
sentence-transformers>=2.7.0
